salesone-backend/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.14,<3.17
djangorestframework-simplejwt>=5.0,<5.3
django-cors-headers>=4.0,<4.8
django-filter>=23.0,<24.0
psycopg2-binary>=2.9,<3.0
celery>=5.3,<5.4
redis>=5.0,<5.1
python-decouple>=3.8,<3.9
gunicorn>=21.2,<22.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0 